requirements.txt
# --- Core Framework ---
Django>=5.0,<6.0
djangorestframework>=3.14

# --- Security / 2FA ---
django-two-factor-auth>=1.18
django-otp>=1.2
django-formtools>=2.5
django-phonenumber-field>=8.0
qrcode>=7.0
phonenumbers>=8.0   # implicit dependency but required

# --- Numerical Computing ---
numpy>=1.26
scipy>=1.11

# --- CORS (dev + frontend integration) ---
django-cors-headers>=4.0

# --- Optional Utilities ---
python-dotenv>=1.0

# --- Optional: image support (Django admin, QR codes, etc.) ---
Pillow>=9.0
